+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## UNRELEASE
+- **Add** : Fixed product customization incrementation error
 
 ## 2.27.5